refactor(runtime): split model provider internals
